Upgrade to Laravel 5.8

+++ database/migrations/2018_10_19_095931_create_tasks_table.php
@@ -14,9 +14,9 @@ class CreateTasksTable extends Migration
     public function up()
     {
         Schema::create('tasks', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
-            $table->integer('task_list_id')->unsigned()->index()->nullable();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->index();
+            $table->bigInteger('task_list_id')->unsigned()->index()->nullable();
             $table->string('task');
             $table->timestamps();
